requirements.txt
langchain
openai
langchain-openai
requests
gql
python-dotenv
fastapi
uvicorn
PyGithub